LPS-135551 Simplify, let's not add this unless we really need it
