+++ modules/apps/object/object-service/src/main/java/com/liferay/object/internal/scope/SiteObjectScopeProvider.java
@@ -15,14 +15,16 @@
 package com.liferay.object.internal.scope;
 
 import com.liferay.application.list.constants.PanelCategoryKeys;
-import com.liferay.object.scope.ObjectScopeGroupHttpRetriever;
 import com.liferay.object.scope.ObjectScopeProvider;
+import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.service.GroupLocalService;
 import com.liferay.portal.kernel.util.Portal;
 
 import java.util.Locale;
+
+import javax.servlet.http.HttpServletRequest;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -37,6 +39,13 @@ import org.osgi.service.component.annotations.Reference;
 public class SiteObjectScopeProvider implements ObjectScopeProvider {
 
 	@Override
+	public long getGroupId(HttpServletRequest httpServletRequest)
+		throws PortalException {
+
+		return _portal.getScopeGroupId(httpServletRequest);
+	}
+
+	@Override
 	public String getKey() {
 		return "site";
 	}
@@ -44,11 +53,6 @@ public class SiteObjectScopeProvider implements ObjectScopeProvider {
 	@Override
 	public String getLabel(Locale locale) {
 		return LanguageUtil.get(locale, "site");
-	}
-
-	@Override
-	public ObjectScopeGroupHttpRetriever getObjectScopeGroupHttpRetriever() {
-		return new PortalObjectScopeGroupHttpRetriever(_portal);
 	}
 
 	@Override
